Loosing HD?

+++ functions/commands/logout.php
@@ -3,7 +3,9 @@ function logout(DbManager $dbManager) {
     if (!$dbManager->transactionBegun) {
         $dbManager->beginTransaction();
     }
-    $dbManager->query('DELETE FROM tokens WHERE token=?', array($_SESSION['token']), 's');
+	if (isset($_SESSION['token'])) {
+		$dbManager->query('DELETE FROM tokens WHERE token=?', array($_SESSION['token']), 's');
+	}
     if (!API_REQUEST) {
         unset($_COOKIE['token']);
         unset($_COOKIE['remember-user']);
